Extract the avatar icon naming, span, and bounds core into a shared module
